requirements.txt
albumentations==1.3.1
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
h5py==3.8.0
numpy==1.21.6
pandas==1.3.5
Pillow==9.5.0
scikit-image==0.19.3
scikit-learn==1.0.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.30.2

